feat: add password reset
get the email info from frontend, validate it and then send the custom email to the user in specified language, which was sent from frontend along with the email. when the user enters new password, validate password, password confirmation, email and token and if everything is right, change the password.